Add files via upload
The code uses the dataset LC25000 Lung and colon histopathological image dataset (view and download here), for the purpose of hyperparameter tuning. It includes 25,000 color images with 5 classes of 5,000 images each. All images are 768 x 768 pixels and are in jpeg file format.

My code still has a problem with indexes. It runs and starts Epochs when the batch size is 1? but when I enter realistic numbers I get a problem. It took me about 2 weeks to create this code!!! Since I have no experience in this field, I apologize for taking such long time and for any mistakes in the code. I merged some of your base code to it. It is also the first time I work on GitHub and create a branch, that's why I had to work isolated on my laptop to not spoil your generated codes!!!! Nonetheless, I continue to work on improving this code.

Please let me know if you have any comments....